basic app conversion
